Fix tests, move GenericFragmentSimulator to Generators subdir
