+++ test/ArtModules/daq_flow_t.cc
@@ -4,7 +4,7 @@
 
 #include "artdaq/Application/LoadParameterSet.hh"
 #include "artdaq/ArtModules/detail/ArtConfig.hh"
-#include "artdaq/DAQdata/GenericFragmentSimulator.hh"
+#include "artdaq/Generators/GenericFragmentSimulator.hh"
 #include "artdaq/DAQrate/SharedMemoryEventManager.hh"
 #include "cetlib_except/exception.h"
 #include "fhiclcpp/ParameterSet.h"
