+++ tools/fcl/genToArt_t.fcl
@@ -26,6 +26,7 @@ process_name: TEST
 
 services: {
     ArtdaqSharedMemoryInterface: { service_provider: ArtdaqSharedMemoryService }
+    ArtdaqFragmentNamingInterface: { service_provider: ArtdaqFragmentNamingService }
 }
 
 source:
